+++ wip-infrastructure/src/main/java/com/cvte/scm/wip/infrastructure/requirement/mapper/dataobject/WipReqLineDO.java
@@ -91,14 +91,14 @@ public class WipReqLineDO {
     @Column(name = "req_qty")
     @ApiModelProperty(value = "需求数量")
     @DecimalMin(value = "0", message = "需求数量必须大于等于零")
-    private Integer reqQty;
+    private Long reqQty;
     /**
      * 已发料数量
      */
     @Column(name = "issued_qty")
     @ApiModelProperty(value = "已发料数量")
     @DecimalMin(value = "0", message = "已发料数量必须大于等于零")
-    private Integer issuedQty;
+    private Long issuedQty;
     /**
      * 超发原因
      */
